Splitting classes into separate files
