+++ src/System.Buffers.Primitives/System/Buffers/OwnedBuffer.cs
@@ -220,60 +220,4 @@ namespace System.Buffers
             return new Internal.OwnedArray<T>(segment);
         }
     }
-
-    interface IKnown
-    {
-        void AddReference();
-        void Release();
-    }
-
-    public struct DisposableReservation<T> : IDisposable
-    {
-        OwnedBuffer<T> _owner;
-
-        internal DisposableReservation(OwnedBuffer<T> owner)
-        {
-            _owner = owner;
-            switch(ReferenceCountingSettings.OwnedMemory) {
-                case ReferenceCountingMethod.Interlocked:
-                    ((IKnown)_owner).AddReference();
-                    break;
-                case ReferenceCountingMethod.ReferenceCounter:
-                    ReferenceCounter.AddReference(_owner);
-                    break;
-                case ReferenceCountingMethod.None:
-                    break;
-            }
-        }
-
-        public Span<T> Span => _owner.Span;
-
-        public void Dispose()
-        {
-            switch (ReferenceCountingSettings.OwnedMemory) {
-                case ReferenceCountingMethod.Interlocked:
-                    ((IKnown)_owner).Release();
-                    break;
-                case ReferenceCountingMethod.ReferenceCounter:
-                    ReferenceCounter.Release(_owner);
-                    break;
-                case ReferenceCountingMethod.None:
-                    break;
-            }
-            _owner = null;
-        }
-    }
-}
-
-namespace System.Runtime
-{
-    public enum ReferenceCountingMethod {
-        Interlocked,
-        ReferenceCounter,
-        None
-    };
-
-    public class ReferenceCountingSettings {
-        public static ReferenceCountingMethod OwnedMemory = ReferenceCountingMethod.Interlocked;
-    }
 }
